Dirty restructuring of project with CMake
