+++ src/CMakeLists.txt
@@ -1,0 +1,6 @@
+cmake_minimum_required(VERSION 3.11)
+project(cbloxfull)
+
+include(GNUInstallDirs)
+
+add_subdirectory(cblox)
