+++ src/cblox/CMakeLists.txt
@@ -1,0 +1,26 @@
+project(cblox)
+
+add_library(cblox
+        src/group.cpp
+        src/http.cpp
+        src/other.cpp
+        src/user.cpp)
+
+target_include_directories(cblox PUBLIC
+        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+        $<INSTALL_INTERFACE:include>
+        PRIVATE src)
+
+target_compile_features(cblox
+        PUBLIC cxx_auto_type
+        PRIVATE cxx_variadic_templates)
+
+target_link_libraries(cblox)
+
+install(TARGETS cblox EXPORT CbloxConfig
+        ARCHIVE  DESTINATION ${CMAKE_INSTALL_LIBDIR}
+        LIBRARY  DESTINATION ${CMAKE_INSTALL_LIBDIR}
+        RUNTIME  DESTINATION ${CMAKE_INSTALL_BINDIR})  # This is for Windows
+install(DIRECTORY include/ DESTINATION ${CMAKE_INSTALL_INCLUDEDIR})
+install(EXPORT CbloxConfig DESTINATION cmake)
+export(TARGETS cblox FILE CbloxConfig.cmake)
